Primary player camera and course repeater
Completed primary playing camera and course repeater coding. Used the Mathf.clamp function to limit the rotation of the camera, so as to avoid warping or losing control of the camera.

For the course repeater, used the rotation.eulerAngels readout to diaply the actual degree rotation of the player game object.