Instituted a basic buoyancy system
For the buoyancy system, just used the y position of the ocean plane, and the y position of the player, and applied upward force to the players' hull. The degree of force was dependent (among other things) on how far into the water the players' hull was. The deeper it was, the more upwards force applied.

ALso tried to switch the rotation of the player to using an AddTorque function. After coding it, the torque was applied along (what looked like) the z axis, instead of the y axis. I swtiched the rotation means back to a simple transforming of the player hull rigidbody.